refactor: extract & extract
